refactor(plugin): extract output_writer.rs + write_extern_err helper (#1079)
Moves the OutputWriter callbacks + HOST_OUTPUT_WRITER_VTABLE +
host_output_writer_vtable() accessor + tier-1 test module out of
host_services/mod.rs into a sibling submodule. mod.rs re-exports
HOST_OUTPUT_WRITER_VTABLE and host_output_writer_vtable (visibility
widened from file-private `static` to `pub static` since runtime_facing
captures it by reference across the submodule boundary).

`write_extern_err` — used by both OutputWriter and InputMailboxes
callbacks — moves to shared/wire.rs so both submodules can call it
after extraction.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
